requirements.txt
PyQt6>=6.5.0
psutil>=5.9.0
requests>=2.31.0

# LangChain AI agent stack
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
langchain-openai>=0.1.0
langsmith>=0.1.75
langgraph>=0.1.0
python-dotenv>=1.0.0
duckduckgo-search>=5.3.0
